+++ src/ViewInstance.js
@@ -21,7 +21,11 @@ import {
   Switch
 } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import {
+  FormattedDate,
+  FormattedTime,
+  FormattedMessage,
+} from 'react-intl';
 
 import queryString from 'query-string';
 
@@ -351,7 +355,7 @@ class ViewInstance extends React.Component {
           </Button>
         </IfPermission>
         <IfPermission perm="ui-inventory.instance.view">
-          { isSourceMARC &&
+          {isSourceMARC &&
             <Button
               id="clickable-view-source"
               buttonStyle="dropdownItem"
@@ -710,6 +714,14 @@ class ViewInstance extends React.Component {
       acc10: areAllFieldsEmpty(Object.values(instanceRelationship)),
     };
 
+    const formattedStatusUpdatedDate = (
+      <React.Fragment>
+        <FormattedDate value={checkIfElementIsEmpty(instanceData.instanceStatusUpdatedDate)} />
+        <br />
+        <FormattedTime value={checkIfElementIsEmpty(instanceData.instanceStatusUpdatedDate)} />
+      </React.Fragment>
+    );
+
     return (
       <Pane
         data-test-instance-details
@@ -795,7 +807,7 @@ class ViewInstance extends React.Component {
                   )}
                 />
                 <Route
-                  path="/inventory/view/"
+                  path="/inventory/(holdings|items|instances)?/view/"
                   render={() => (
                     <this.cHoldings
                       dataKey={id}
@@ -879,7 +891,7 @@ class ViewInstance extends React.Component {
             <Col xs={3}>
               <KeyValue
                 label={<FormattedMessage id="ui-inventory.instanceStatusUpdatedDate" />}
-                value={checkIfElementIsEmpty(instanceData.instanceStatusUpdatedDate)}
+                value={formattedStatusUpdatedDate}
               />
             </Col>
           </Row>