+++ test/bigtest/tests/item-view-page-test.js
@@ -45,6 +45,18 @@ describe('ItemViewPage', () => {
         expect(ItemViewPage.enumerationDataAccordion.isOpen).to.be.false;
       });
 
+      describe('"Administrative data" section', () => {
+        it('contains "Statistical code" table', () => {
+          expect(ItemViewPage.statisticalCodeTable.isPresent).to.be.true;
+        });
+      });
+
+      describe('"Electronic access" section', () => {
+        it('contains "Electronic access" table', () => {
+          expect(ItemViewPage.electronicAccessTable.isPresent).to.be.true;
+        });
+      });
+
       describe('clicking on edit button in the pane header', () => {
         beforeEach(async () => {
           await ItemViewPage.clickEditItemButton();